requirements.txt
# ── Core async / web ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0

# ── AI / LLM ─────────────────────────────────────────────────────────────────
openai>=1.30.0
openai-agents>=0.0.6

# ── Speech-to-text ───────────────────────────────────────────────────────────
faster-whisper>=1.0.1

# ── Wake word ────────────────────────────────────────────────────────────────
pvporcupine>=3.0.0

# ── Voice Activity Detection ─────────────────────────────────────────────────
silero-vad>=5.0
torch>=2.2.0
torchaudio>=2.2.0

# ── Speaker verification ─────────────────────────────────────────────────────
speechbrain>=1.0.0

# ── TTS ──────────────────────────────────────────────────────────────────────
elevenlabs>=1.2.0
pywin32>=306

# ── Audio I/O ────────────────────────────────────────────────────────────────
pyaudio>=0.2.14
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0

# ── Browser automation ───────────────────────────────────────────────────────
playwright>=1.44.0

# ── Screen capture & OCR ─────────────────────────────────────────────────────
mss>=9.0.1
pytesseract>=0.3.10
Pillow>=10.3.0

# ── Storage ──────────────────────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Google APIs (Gmail + Calendar) ───────────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.128.0

# ── Document processing ───────────────────────────────────────────────────────
python-docx>=1.1.0
openpyxl>=3.1.2
PyPDF2>=3.0.1

# ── Clipboard ─────────────────────────────────────────────────────────────────
pyperclip>=1.8.2

# ── CUA (visual automation) ───────────────────────────────────────────────────
pyautogui>=0.9.54

# ── Utilities ────────────────────────────────────────────────────────────────
rich>=13.7.0
typer>=0.12.0
pynput>=1.7.6
psutil>=5.9.8
keyboard>=0.13.5

# ── System tray icon (Phase 4) ────────────────────────────────────────────────
pystray>=0.19.5
# Pillow already listed above (>=10.3.0)
